feat: add support for git repository detection
When running the command and not specifying a direct repository to
operate on via the `-r,--repo` argument, the CLI now tries to retrieve
the repository URL by first checking if the current directory is a valid
git repository and then trying to parse its `origin` remote.

The CLI argument will take precedence over the automatic detection
process.